+++ src/apps/draft-service.ts
@@ -144,6 +144,25 @@ async function generateLocalFormSubmissionDraftsFromStorage(
       deletedDraftIds,
     )
 
+  async function broadcastUpdate() {
+    await executeDraftsListeners(
+      Array.from(localFormSubmissionDraftsMap.values()),
+    )
+  }
+
+  // At this point we need to store the state of the drafts in localForage
+  const username = getUsername()
+  if (!username) {
+    throw new OneBlinkAppsError(
+      'You cannot download drafts until you have logged in. Please login and try again.',
+      {
+        requiresLogin: true,
+      },
+    )
+  }
+
+  const draftsToDownload: SubmissionTypes.FormSubmissionDraft[] = []
+
   for (const formSubmissionDraft of localDraftsStorage.syncedFormSubmissionDrafts) {
     if (
       // Unsycned version of draft takes priority over the synced version
@@ -153,6 +172,33 @@ async function generateLocalFormSubmissionDraftsFromStorage(
       // Remove any drafts deleted while offline
       !deletedDraftIds.has(formSubmissionDraft.id)
     ) {
+      const localDraftSubmission = await getLocalDraftSubmission(
+        formSubmissionDraft.id,
+      )
+      draftsToDownload.push(formSubmissionDraft)
+      localFormSubmissionDraftsMap.set(formSubmissionDraft.id, {
+        ...formSubmissionDraft,
+        downloadStatus: 'PENDING',
+        draftSubmission: localDraftSubmission ?? undefined,
+      })
+    }
+  }
+
+  await broadcastUpdate()
+
+  // TODO Batch the downloads instead of sequentially
+  if (draftsToDownload.length) {
+    for (const formSubmissionDraft of draftsToDownload) {
+      const currentValue = localFormSubmissionDraftsMap.get(
+        formSubmissionDraft.id,
+      )
+      if (currentValue) {
+        localFormSubmissionDraftsMap.set(formSubmissionDraft.id, {
+          ...currentValue,
+          downloadStatus: 'DOWNLOADING',
+        })
+      }
+      await broadcastUpdate()
       const draftSubmission = await getDraftSubmission(
         formSubmissionDraft,
       ).catch((err) => {
@@ -164,8 +210,10 @@ async function generateLocalFormSubmissionDraftsFromStorage(
       })
       localFormSubmissionDraftsMap.set(formSubmissionDraft.id, {
         ...formSubmissionDraft,
-        draftSubmission,
+        draftSubmission: draftSubmission,
+        downloadStatus: draftSubmission ? 'SUCCESS' : 'ERROR',
       })
+      await broadcastUpdate()
     }
   }
 
@@ -800,24 +848,26 @@ async function syncDrafts({
       localDraftsStorage.syncedFormSubmissionDrafts = formSubmissionDrafts
     }
 
-    await setAndBroadcastDrafts(localDraftsStorage)
-
-    if (localDraftsStorage.syncedFormSubmissionDrafts.length) {
-      console.log(
-        'Ensuring all draft data is available for offline use for synced drafts',
-        localDraftsStorage.syncedFormSubmissionDrafts,
-      )
-      for (const formSubmissionDraft of localDraftsStorage.syncedFormSubmissionDrafts) {
-        await getDraftSubmission(formSubmissionDraft, abortSignal).catch(
-          (error) => {
-            console.warn('Could not download Draft Data as JSON', error)
-          },
+    console.log('Ddownloading drafts in the background')
+    setAndBroadcastDrafts(localDraftsStorage).then(async () => {
+      if (localDraftsStorage.syncedFormSubmissionDrafts.length) {
+        console.log(
+          'Ensuring all draft data is available for offline use for synced drafts',
+          localDraftsStorage.syncedFormSubmissionDrafts,
         )
+        for (const formSubmissionDraft of localDraftsStorage.syncedFormSubmissionDrafts) {
+          await getDraftSubmission(formSubmissionDraft, abortSignal).catch(
+            (error) => {
+              console.warn('Could not download Draft Data as JSON', error)
+            },
+          )
+        }
       }
-    }
+      console.log('Finished syncing drafts.')
+      _isSyncingDrafts = false
+    })
 
-    console.log('Finished syncing drafts.')
-    _isSyncingDrafts = false
+    // broadcast the drafts and download the draft data in the background
   } catch (error) {
     _isSyncingDrafts = false
     if (abortSignal?.aborted) {