+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - github action permissions to allow docs and github release jobs
+- Lot / DP Numbers display for NSW Point V3 form element
 
 ### Dependencies