+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Lot / DP Numbers display for NSW Point V3 form element
+- disabled on `Clickable` not disabling click function
 
 ## [9.1.0] - 2026-02-23